compiler: Move SymbolRegistry to ir/
