+++ devito/ir/support/__init__.py
@@ -5,3 +5,4 @@ from .guards import *  # noqa
 from .syncs import *  # noqa
 from .utils import *  # noqa
 from .properties import *  # noqa
+from .symregistry import *  # noqa
