+++ devito/ir/support/symregistry.py
