+++ devito/operator/__init__.py
@@ -1,4 +1,3 @@
-from .symbols import SymbolRegistry  # noqa
 from .operator import Operator  # noqa
 from .profiling import profiler_registry  # noqa
 from .registry import operator_registry  # noqa
